** Refactor ** Split logic, creating SolitaireDice class
